+++ src/main/java/de/samply/db/model/Query.java
@@ -53,4 +53,7 @@ public class Query {
     @Column(name = "explorer_url")
     private String explorerUrl;
 
+    @Column(name = "context")
+    private String context;
+
 }