+++ CHANGELOG.md
@@ -58,3 +58,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Download application form
 - Edit project and query
 - Fetch output format, query format and templates
+- Query Context